add renderComments.js
